Monte Carlo for big num

+++ exercise04/solution/montecarlo_seq.cpp
@@ -11,7 +11,7 @@
 
 int main(int argc, char** argv) {
 
-  long s = N;
+  long long s = N;
   char* temp;
   if (argc >= 2) {
     s = strtoul(argv[1],&temp,0);
@@ -20,11 +20,11 @@ int main(int argc, char** argv) {
     // calculate random points
     srandom(time(NULL));
 
-  long hit = 0;
+  long long hit = 0;
   VALUE x;
   VALUE y;
   // Check if it is in circle
-  for (int i=0; i<s; i++) {
+  for (long long i=0; i<s; i++) {
     x = (VALUE)rand()/RAND_MAX;
     y = (VALUE)rand()/RAND_MAX;
     VALUE temp = pow(x,2)+pow(y,2);
